Add reactToNotification

+++ src/main/java/com/safetouch/dal/repositories/NotificationRepository.java
@@ -1,12 +1,13 @@
 package com.safetouch.dal.repositories;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.safetouch.dal.entities.Notification;
 
-public interface NotificationRepository extends JpaRepository<Notification, Long> {
+public interface NotificationRepository extends JpaRepository<Notification, BigInteger> {
 
 	List<Notification> findByNotifiedAndAdminEmail(boolean notified, String email);
